+++ Models/Publisher.cs
@@ -9,12 +9,12 @@ namespace MVCPlayWithMe.Models
     {
         // Chiết khấu so với giá bìa khi nhập sách.
         // Dùng để tham khảo khi tính giá bán thực tế.Giá trị mặc định là 20
-        public int discount { get; set; }
+        public float discount { get; set; }
 
         public string detail { get; set; }
 
         public Publisher(int idInput, string nameInput,
-            int discountInput, string detalInput) : base(idInput, nameInput)
+            float discountInput, string detalInput) : base(idInput, nameInput)
         {
             discount = discountInput;
             detail = detalInput;
